refactor: change request error handling
