--- graphiql-app/src/components/DraggableDiv/index.ts
@@ -1,1 +1,0 @@
-export { default } from './DraggableDiv'
