+++ graphiql-app/src/components/DraggableElement/DraggableElement.module.scss
