+++ graphiql-app/src/components/DraggableElement/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DraggableElement'
